producer/requirements.txt
kafka-python==2.0.2
pandas==2.1.0
numpy==1.24.0
python-dotenv==1.0.0
prometheus-client==0.17.1
tqdm==4.66.1